Update internal objects to match the API
