+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmResultsImpl.kt
@@ -22,6 +22,8 @@ import io.realm.internal.interop.Callback
 import io.realm.internal.interop.NativePointer
 import io.realm.internal.interop.RealmCoreException
 import io.realm.internal.interop.RealmInterop
+import io.realm.notifications.InitialListImpl
+import io.realm.notifications.ListChange
 import kotlinx.coroutines.channels.ChannelResult
 import kotlinx.coroutines.channels.SendChannel
 import kotlinx.coroutines.flow.Flow
@@ -39,7 +41,7 @@ internal class RealmResultsImpl<E : RealmObject> constructor(
     private val clazz: KClass<E>,
     private val mediator: Mediator,
     private val mode: Mode = Mode.RESULTS
-) : AbstractList<E>(), RealmResults<E>, Observable<RealmResultsImpl<E>, RealmResultsImpl<E>>, RealmStateHolder {
+) : AbstractList<E>(), RealmResults<E>, Observable<RealmResultsImpl<E>, ListChange<RealmResultsImpl<E>>>, RealmStateHolder, Flowable<ListChange<RealmResults<E>>> {
 
     enum class Mode {
         // FIXME Needed to make working with @LinkingObjects easier.
@@ -70,7 +72,7 @@ internal class RealmResultsImpl<E : RealmObject> constructor(
         }
     }
 
-    override fun asFlow(): Flow<RealmResultsImpl<E>> {
+    override fun asFlow(): Flow<ListChange<RealmResults<E>>> {
         realm.checkClosed()
         return realm.owner.registerObserver(this)
     }
@@ -108,10 +110,17 @@ internal class RealmResultsImpl<E : RealmObject> constructor(
     override fun emitFrozenUpdate(
         frozenRealm: RealmReference,
         change: NativePointer,
-        channel: SendChannel<RealmResultsImpl<E>>
+        channel: SendChannel<ListChange<RealmResultsImpl<E>>>
     ): ChannelResult<Unit>? {
         val frozenResult = freeze(frozenRealm)
-        return channel.trySend(frozenResult)
+
+        val builder = UpdatedListBuilderImpl(frozenResult, change)
+
+        return if (builder.isEmpty()) {
+            channel.trySend(InitialListImpl(frozenResult))
+        } else {
+            channel.trySend(builder.build())
+        }
     }
 
     override fun realmState(): RealmState = realm
